requirements.txt
pygame==2.5.2
python-socketio==5.9.0
requests==2.31.0
flask